+++ CHANGELOG.md
@@ -69,6 +69,7 @@ Compatibility:
 * Fix `Module#module_function` and keep new Module methods public created from implicitly private callbacks (#4388, @andrykonchin).
 * Add `flags` keyword argument to `Dir.glob` as an alternative to the positional argument (#4394, @earlopain).
 * Fix `new` on `Data` subclasses to pass the given block to `#initialize` (@eregon).
+* Fix `ENV` for `BINARY` strings with non-ASCII bytes (@nirvdrum).
 
 Performance: